Break daemon lifecycle↔shutdown-handlers import cycle
lifecycle.ts imported installShutdownHandlers from shutdown-handlers.ts, and
shutdown-handlers.ts imported stopDiskPressureGuardForLifecycle back from
lifecycle.ts — a mutual value import (real runtime coupling, not a type-only
edge).

Move the disk-pressure guard lifecycle wiring
(start/stopDiskPressureGuardForLifecycle plus the deferred startup-sample timer
and its state) out of lifecycle.ts into a new leaf module
daemon/disk-pressure-guard-lifecycle.ts. Both lifecycle.ts and
shutdown-handlers.ts now import from that leaf, so shutdown-handlers no longer
imports lifecycle and the cycle is gone. The functions are self-contained
around disk-pressure-guard.ts, so lifecycle.ts drops those imports entirely.

Also address review feedback carried over from #37324: startEmbeddingRuntimeManager
is now a plain async function instead of wrapping its body in a redundant
`void (async () => { … })()` IIFE; the lifecycle caller voids the returned
promise to keep the fire-and-forget semantics.

madge circular chains: 415 → 414; no cycle now touches lifecycle.ts,
shutdown-handlers.ts, or the new module.
