+++ assistant/src/daemon/lifecycle.ts
@@ -71,11 +71,7 @@ import {
   setDbReady,
   setStartupComplete,
 } from "./daemon-readiness.js";
-import {
-  evaluateDiskPressureNow,
-  startDiskPressureGuard,
-  stopDiskPressureGuard,
-} from "./disk-pressure-guard.js";
+import { startDiskPressureGuardForLifecycle } from "./disk-pressure-guard-lifecycle.js";
 import { startEventLoopWatchdog } from "./event-loop-watchdog.js";
 import { initializePlugins } from "./external-plugins-bootstrap.js";
 import { backfillSlackInjectionTemplates } from "./handlers/config-slack-channel.js";
@@ -91,57 +87,9 @@ import { installShutdownHandlers } from "./shutdown-handlers.js";
 import { broadcastDaemonStatus } from "./status.js";
 
 const log = getLogger("lifecycle");
-let diskPressureStartupSampleTimer: ReturnType<typeof setTimeout> | null = null;
 
 function loadDotEnv(): void {
   dotenvConfig({ path: getDotEnvPath(), quiet: true });
-}
-
-function runDeferredDiskPressureStartupSample(): void {
-  diskPressureStartupSampleTimer = null;
-  try {
-    const status = evaluateDiskPressureNow();
-    if (status.error) {
-      log.warn(
-        { error: status.error },
-        "Disk pressure guard sample failed during startup — continuing unlocked",
-      );
-    }
-  } catch (err) {
-    log.warn(
-      { err },
-      "Disk pressure guard failed during startup — continuing unlocked",
-    );
-  }
-}
-
-export function startDiskPressureGuardForLifecycle(): void {
-  try {
-    const startedStatus = startDiskPressureGuard();
-    if (!startedStatus.enabled) {
-      return;
-    }
-    if (!diskPressureStartupSampleTimer) {
-      diskPressureStartupSampleTimer = setTimeout(
-        runDeferredDiskPressureStartupSample,
-        0,
-      );
-      (diskPressureStartupSampleTimer as { unref?: () => void }).unref?.();
-    }
-  } catch (err) {
-    log.warn(
-      { err },
-      "Disk pressure guard failed during startup — continuing unlocked",
-    );
-  }
-}
-
-export function stopDiskPressureGuardForLifecycle(): void {
-  if (diskPressureStartupSampleTimer) {
-    clearTimeout(diskPressureStartupSampleTimer);
-    diskPressureStartupSampleTimer = null;
-  }
-  stopDiskPressureGuard();
 }
 
 // Entry point for the daemon process itself
@@ -719,7 +667,7 @@ export async function runDaemon(): Promise<void> {
   // failure never affects startup.
   installAssistantSymlink();
 
-  startEmbeddingRuntimeManager();
+  void startEmbeddingRuntimeManager();
 
   startWorkspaceHeartbeatService();
